Fix failing master (#115)
* Add black formatting

* Fix some of the failing tests

* Fix ZipF by specifying dtypes

* Simply changes to zipf

* Fix Bernoulli types

* Change equality checks

* Revert numpy coercion

* Add black formatting

* Fix categorical distribution test

* Fix linting errors

+++ pymc4/random_variables/discrete.py
@@ -60,7 +60,7 @@ class Bernoulli(RandomVariable):
     """
 
     def _base_dist(self, p: TensorLike, *args, **kwargs):
-        return tfd.Bernoulli(probs=p, *args, **kwargs)
+        return tfd.Bernoulli(probs=p, dtype=tf.float32, *args, **kwargs)
 
 
 class Binomial(RandomVariable):
@@ -219,7 +219,7 @@ class Categorical(RandomVariable):
     """
 
     def _base_dist(self, p: TensorLike, *args, **kwargs):
-        return tfd.Categorical(probs=p, *args, **kwargs)
+        return tfd.Categorical(probs=p, dtype=tf.float32, *args, **kwargs)
 
 
 class Geometric(RandomVariable):
@@ -637,4 +637,4 @@ class Zipf(RandomVariable):
     """
 
     def _base_dist(self, alpha: TensorLike, *args, **kwargs):
-        return tfd.Zipf(power=alpha, *args, **kwargs)
+        return tfd.Zipf(power=alpha, dtype=tf.float32, *args, **kwargs)
